+++ modules/betterchairs-plugin/src/main/java/de/sprax2013/betterchairs/Updater.java
@@ -25,6 +25,8 @@ import java.util.TimerTask;
 public class Updater implements Listener {
     public static final String VERSION_TXT_URL = "https://spraxdev.github.io/BetterChairs/version.txt";
 
+    public static final String MODRINTH_URL = "https://modrinth.com/plugin/betterchairs";
+    public static final String HANGAR_URL = "https://hangar.papermc.io/Sprax/BetterChairs";
     public static final String SPIGOT_MC_URL = "https://r.spiget.org/84809";
     public static final String CRAFTARO_URL = "https://craftaro.com/marketplace/product/489";
     public static final String GITHUB_URL = "https://github.com/SpraxDev/BetterChairs/releases";
@@ -95,8 +97,8 @@ public class Updater implements Listener {
             this.newerVersion = versionStr;
 
             ChairManager.getLogger()
-                    .info(() -> String.format("Found a new update v%s -> v%s — Download the update from: %nSpigotMC: %s%nCraftaro: %s%nGitHub: %s%n%nChangelog: %s",
-                            currVersion, versionStr, SPIGOT_MC_URL, CRAFTARO_URL, GITHUB_URL, getChangelogUrl(versionStr)));
+                    .info(() -> String.format("Found a new update v%s -> v%s — Download the update from: %nmodrinth: %s%nHangar: %s%nSpigotMC: %s%nCraftaro: %s%nGitHub: %s%n%nChangelog: %s",
+                            currVersion, versionStr, MODRINTH_URL, HANGAR_URL, SPIGOT_MC_URL, CRAFTARO_URL, GITHUB_URL, getChangelogUrl(versionStr)));
         } else {
             this.newerVersion = null;
         }
@@ -115,6 +117,14 @@ public class Updater implements Listener {
                         .append("] ").color(ChatColor.GRAY)
                         .append("Found a new update v" +
                                 this.plugin.getDescription().getVersion() + " -> v" + this.newerVersion).color(ChatColor.YELLOW)
+                        .append(" [modrinth] ")
+                        .color(ChatColor.GREEN)
+                        .event(new ClickEvent(ClickEvent.Action.OPEN_URL, MODRINTH_URL))
+                        .event(new HoverEvent(HoverEvent.Action.SHOW_TEXT, TextComponent.fromLegacyText("§2Click to visit the download page on modrinth")))
+                        .append(" [Hangar] ")
+                        .color(ChatColor.GREEN)
+                        .event(new ClickEvent(ClickEvent.Action.OPEN_URL, HANGAR_URL))
+                        .event(new HoverEvent(HoverEvent.Action.SHOW_TEXT, TextComponent.fromLegacyText("§2Click to visit the download page on Hangar")))
                         .append(" [SpigotMC] ")
                         .color(ChatColor.GREEN)
                         .event(new ClickEvent(ClickEvent.Action.OPEN_URL, SPIGOT_MC_URL))